requirements.txt
absl-py==1.4.0
albumentations==1.3.1
altair==5.1.2
altgraph==0.17.4
annotated-types==0.6.0
annoy==1.17.3
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.16.0
comm==0.1.4
confection==0.1.3
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.8
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
efficientnet-pytorch==0.7.1
executing==2.0.0
faiss-cpu==1.7.4
fastai==2.7.13
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema==2.18.1
fastprogress==1.0.3
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.42.0
fqdn==1.5.1
fsspec==2023.9.1
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
huggingface-hub==0.17.2
idna==3.4
imageio==2.31.3
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==2.13.1
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.6
macholib==1.16.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
munch==4.0.0
murmurhash==1.0.10
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.5
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.11.0
preshed==3.0.9
pretrainedmodels==0.7.4
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
qudida==0.0.4
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
segmentation-models-pytorch==0.3.3
Send2Trash==1.8.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.28.1
sympy==1.12
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
tensorflow-recommenders==0.7.3
termcolor==2.3.0
terminado==0.17.1
thinc==8.2.1
threadpoolctl==3.2.0
tifffile==2023.9.18
timm==0.9.2
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
urllib3==1.26.16
validators==0.22.0
wasabi==1.1.2
wcwidth==0.2.8
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.3.7
widgetsnbextension==4.0.9
wrapt==1.15.0
zipp==3.17.0
